Make Blobby death idempotent

+++ ascend/mobs.py
@@ -524,9 +524,11 @@ class Blobby:
         )
 
     def delete(self):
-        self.anim.stop()
-        clock.unschedule(self.bounce)
-        self.shape.delete()
+        if self.shape:
+            self.anim.stop()
+            clock.unschedule(self.bounce)
+            self.shape.delete()
+            self.shape = None
 
     def die(self, vel=Vector2D()):
         self.scene.smoke.emit(
